requirements.txt
dj-database-url==3.0.1
django-cors-headers==4.9.0
django-modeltranslation==0.19.17
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Pillow==10.3.0
psycopg2-binary==2.9.10
requests==2.32.3
sentry-sdk==2.40.0
whitenoise==6.11.0
